src/requirements.txt
requests>=2.31.0
pyjwt>=2.8.0
